Fox type on ProviderFeedbackRequestHistory Status

+++ src/Sfa.Tl.Matching.Application.UnitTests/Services/ProviderFeedback/When_ProviderFeedbackService_Is_Called_To_Send_Provider_Quarterly_Update_Emails.cs
@@ -72,7 +72,7 @@ namespace Sfa.Tl.Matching.Application.UnitTests.Services.ProviderFeedback
                 .Received(1)
                 .Update(Arg.Is<ProviderFeedbackRequestHistory>(
                     p => p.Id == 1
-                         && p.Status == (int)ProviderFeedbackRequestStatus.Sent
+                         && p.Status == (short)ProviderFeedbackRequestStatus.Sent
                          && p.ProviderCount == 1
                          && p.ModifiedBy == "TestUser"
                          ));
